backend/requirements.txt
Flask==2.3.3
openai==1.30.1
elevenlabs==1.1.0
requests==2.31.0
python-dotenv==1.0.1
google-generativeai
Werkzeug==2.3.7
python-docx
replicate
huggingface_hub
pillow
requests
gtts
pydub